+++ src/main.rs
@@ -6,6 +6,7 @@
 #![cfg_attr(target_os = "windows", windows_subsystem = "windows")]
 
 mod audio;
+mod autostart;
 mod hotkey;
 mod logging;
 mod models;